Move the hook an use it in both places
